+++ examples/cityscapes/utils_all/metrics.py
@@ -1,10 +1,15 @@
 from typing import Tuple
+
+import numpy as np
 import tensorflow as tf
 
-from project_config import MODEL_FORMAT, IMAGE_SIZE
+from project_config import MODEL_FORMAT, IMAGE_SIZE, BACKGROUND_LABEL
+from utils_all.general_utils import get_predict_bbox_list, bb_array_to_object
+from utils_all.preprocessing import CATEGORIES_no_background, Cityscapes
 from yolo_helpers.yolo_utils import LOSS_FN
 
 from code_loader.helpers.detection.yolo.utils import reshape_output_list
+from code_loader.helpers.detection.utils import xyxy_to_xywh_format, xywh_to_xyxy_format
 
 def compute_losses(obj_true: tf.Tensor, od_pred: tf.Tensor) -> Tuple[tf.Tensor, tf.Tensor, tf.Tensor]:
     """
@@ -69,4 +74,73 @@ def object_metric(bb_gt: tf.Tensor, detection_pred: tf.Tensor) -> tf.Tensor:
     """
     _, _, loss_o = compute_losses(bb_gt, detection_pred)
     return tf.reduce_sum(loss_o, axis=0)[:, 0]  # shape of batch
+
+def convert_to_xyxy(bounding_boxes):
+    xyxy_boxes = []
+    for box in bounding_boxes:
+        center_x, center_y, width, height, label = box.x, box.y, box.width, box.height, box.label
+        class_id = Cityscapes.get_class_id(label)
+        x_min = (center_x - width / 2) * IMAGE_SIZE[0]
+        y_min = (center_y - height / 2) * IMAGE_SIZE[1]
+        x_max = (center_x + width / 2) * IMAGE_SIZE[0]
+        y_max = (center_y + height / 2) * IMAGE_SIZE[1]
+        xyxy_boxes.append([x_min, y_min, x_max, y_max, class_id])
+    return xyxy_boxes
+
+
+def intersection_area(true_box, pred_box):
+  """Calculates the intersection area between two bounding boxes.
+
+  Args:
+    true_box: A bounding box in the format [x1, y1, x2, y2].
+    pred_box: A bounding box in the format [x1, y1, x2, y2].
+
+  Returns:
+    The intersection area.
+  """
+
+  x1 = max(true_box[0], pred_box[0])
+  y1 = max(true_box[1], pred_box[1])
+  x2 = min(true_box[2], pred_box[2])
+  y2 = min(true_box[3], pred_box[3])
+  if x2 < x1 or y2 < y1:
+    return 0
+  else:
+    return (x2 - x1) * (y2 - y1)
+
+def union_area(true_box, pred_box):
+  """Calculates the union area between two bounding boxes.
+
+  Args:
+    true_box: A bounding box in the format [x1, y1, x2, y2].
+    pred_box: A bounding box in the format [x1, y1, x2, y2].
+
+  Returns:
+    The union area.
+  """
+
+  true_area = (true_box[2] - true_box[0]) * (true_box[3] - true_box[1])
+  pred_area = (pred_box[2] - pred_box[0]) * (pred_box[3] - pred_box[1])
+  return true_area + pred_area - intersection_area(true_box, pred_box)
+
+def calculate_iou(y_true, y_pred):
+  """Calculates the intersection over union (IoU) between a list of true bounding boxes and a list of predicted bounding boxes.
+
+  Args:
+    true_boxes: A list of y bounding boxes in the format [x1, y1, x2, y2].
+    pred_boxes: A list of y bounding boxes in the format [x1, y1, x2, y2].
+
+  Returns:
+    A list of y IoU scores.
+  """
+
+  iou_scores = []
+  for true_box, pred_box in zip(y_true, y_pred):
+    intersection = intersection_area(true_box, pred_box)
+    union = union_area(true_box, pred_box)
+    # Calculate the IOU value
+    iou = tf.where(union > 0, intersection / union, 0)
+    iou_scores.append(iou)
+
+  return iou_scores
